+++ app/backend/src/lib/types/logger.ts
@@ -1,0 +1,12 @@
+
+export type LogType = "all"|"errors"|"warnings"|"infos"|"debug";
+
+export type LogRecord = {
+  uuid: string;
+  type: LogType;
+  timestamp: string;
+  issuer?: string;
+  action?: string;
+  message: string;
+}
+
