+++ app/pages/src/lib/types/notification.ts
@@ -1,5 +1,4 @@
 
-
 export interface Notification {
   id?: string;
   success: boolean;
