+++ app/pages/src/lib/types/step.ts
@@ -1,0 +1,7 @@
+
+import { type ComponentType, type SVGAttributes } from "preact";
+
+export interface StepPair {
+  icon: ComponentType<SVGAttributes<SVGSVGElement>>;
+  label: string;
+}
